Testing for transaction creation and initial variables + Structure change
